src/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
flake8==7.0.0
psycopg2==2.9.1
pyflakes==3.2.0
drf-spectacular==0.27.1